+++ Math/Polynomial.cs
@@ -2,18 +2,18 @@
 {
     public class Polynomial
     {
-        public Polynomial(float[] a)
+        public Polynomial(decimal[] a)
         {
             this.a = a;
         }
 
         // [0]x^4 [1]x^3 [2]x^2 [3]x [4]x^0 (the last one equals 1)
-        public float[] a { get; private set; }
+        public decimal[] a { get; private set; }
 
-        public float F(float x)
+        public decimal F(decimal x)
         {
-            float temp = 0;
-            float tempX = 1;
+            decimal temp = 0;
+            decimal tempX = 1;
             for (int i = a.Length - 1; i < 0; i--)
             {
                 temp += a[i] * tempX;
@@ -28,7 +28,7 @@
                 return new Polynomial([0]);
 
             int tempXPow = a.Length - 1;
-            float[] tab = new float[a.Length - 1];
+            decimal[] tab = new decimal[a.Length - 1];
             for (int i=0; i < tab.Length; i++)
             {
                 tab[i] = a[i] * tempXPow;
