+++ Math/Triangle.cs
@@ -2,13 +2,13 @@
 {
     internal class Triangle
     {
-        public float a { get; private set; }
-        public float b { get; private set; }
-        public float c { get; private set; }
+        public decimal a { get; private set; }
+        public decimal b { get; private set; }
+        public decimal c { get; private set; }
 
-        public float a_h { get; private set; }
-        public float b_h { get; private set; }
-        public float c_h { get; private set; }
+        public decimal a_h { get; private set; }
+        public decimal b_h { get; private set; }
+        public decimal c_h { get; private set; }
 
 
     }
